tenantServer/requirements.txt
Flask==2.1.0
flask-pymongo==2.3.0
pymongo==3.12.0
openai==0.27.0
requests==2.26.0
gTTS==2.2.3
python-dotenv==0.19.1
scikit-learn==0.24.2
langid==1.1.6
bcrypt==3.2.0
PyJWT==2.3.0
Werkzeug==2.0.2
gunicorn==20.1.0